job-hunter/requirements.txt
# Scraping
playwright==1.52.0
playwright-stealth==1.0.6
beautifulsoup4==4.13.4

# IA
anthropic==0.52.0

# Dashboard
flask==3.1.1

# Base de données
# SQLite est inclus dans Python stdlib

# Config
pyyaml==6.0.2

# Utilitaires
requests==2.32.3
